requirement.txt
Django==3.2.11
djangorestframework==3.13.1
django-allauth==0.51.0
djangorestframework==3.14.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-cors-headers==3.13.0
django-rest-auth==0.9.5
